// whitebirrd-ssms-ui/src/app/admin/mbo/models/ObjectiveStatus.ts
export enum ObjectiveStatus {
   // Initiated
   Created = 0,
   // Pending Approval
   Locked = 1,
   Approved = 2,
   Denied = 3,
   // Work in Progress
   WorkInProgress = 4,
   // Pending Review
   InReview = 5,
   // Pending Extension
   Extension = 6,
   // Gave Up
   GaveUp = 7,
   Rejected = 8,
   // Un-Acceptable
   Unacceptable = 9,
   Complete = 10
}